# server/docs/swagger.yaml
basePath: /
definitions:
...
  community.CommentThumbsUp:
    properties:
      commentId:
        description: 评论编号
        type: integer
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      userId:
        description: "用户id \t//点赞时间"
        type: integer
    type: object
  community.ForumComment:
    properties:
      checkStatus:
        description: 审核状态(0未审批 1通过 2拒绝)
        type: integer
      checkTime:
        description: 审核时间
        type: string
      checkUser:
        description: 审核人
        type: integer
      commentContent:
        description: 评论内容
        type: string
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      likeTimes:
        description: 点赞数
        type: integer
      parentId:
        description: 父评论编号
        type: integer
      postsId:
        description: 帖子编号
        type: integer
      userId:
        description: 评论者
        type: integer
    type: object
  community.ForumPosts:
    properties:
      anonymity:
        description: 匿名发布(0否 1是)
        type: integer
      attachment:
        description: 附件
        type: string
      category:
        description: 类别（0视频、1动态、2资讯、3公告、4文章、5问答、6建议）
        type: integer
      checkStatus:
        description: 审核状态(0草稿 1未审批 2通过 3拒绝)
        type: integer
      checkTime:
        description: 审核时间
        type: string
      checkUser:
        description: 审核人
        type: integer
      circleId:
        description: 圈子_编号
        type: integer
      collectNum:
        description: 收藏次数
        type: integer
      commentId:
        description: 问答最佳答案ID
        type: integer
      commentNum:
        description: 评论次数
        type: integer
      contentHtml:
        description: html内容
        type: string
      contentMarkdown:
        description: markdown内容
        type: string
      contentType:
        description: 内容类型:0 markdown,1 html
        type: integer
      coverImage:
        description: 封面
        type: string
      createdAt:
        description: 创建时间
        type: string
      groupId:
        description: 帖子分类编号
        type: integer
      id:
        description: 主键ID
        type: integer
      likeNum:
        description: 点赞次数
        type: integer
      marrow:
        description: 精华(0否 1是)
        type: integer
      pay:
        description: 付费：0 否，1是
        type: integer
      payAttachment:
        description: 附件付费：0 否，1是
        type: integer
      payContent:
        description: 内容付费：0 否，1是
        type: integer
      payContentLook:
        description: 内容付费可查看百分比例
        type: integer
      payCurrency:
        description: 付费货币：0 人民、1积分、2代币
        type: integer
      payNum:
        description: 付费金额
        type: integer
      powerComment:
        description: 评论权限(0关闭 1开启)
        type: integer
      powerCommentAnonymity:
        description: 匿名评论(0关闭 1开启)
        type: integer
      readNum:
        description: 阅读次数
        type: integer
      seoIntroduce:
        description: 简介(SEO简介)
        type: string
      seoKey:
        description: SEO关键词
        type: string
      shareNum:
        description: 分享次数
        type: integer
      source:
        description: 来源
        type: string
      tag:
        description: 标签
        type: string
      time:
        description: 发布时间
        type: string
      title:
        description: 标题
        type: string
      top:
        description: 置顶(0否 1是)
        type: integer
      userId:
        description: 发布者编号
        type: integer
      video:
        description: 视频地址
        type: string
    type: object
...
  community.ForumThumbsUp:
    properties:
      createdAt:
        description: 创建时间
        type: string
      id:
        description: 主键ID
        type: integer
      postsId:
        description: 帖子编号
        type: integer
      userId:
        description: 用户id
        type: integer
    type: object
...
  request.CreateForumComment:
    properties:
      commentContent:
        description: 评论内容
        type: string
      postsId:
        description: 帖子编号
        type: integer
    type: object
...
  request.DeleteCommentThumbsUp:
    properties:
      commentId:
        description: 评论编号
        type: integer
    type: object
  request.DeleteForumThumbsUp:
    properties:
      postsId:
        description: 帖子编号
        type: integer
    type: object
...
  request.IdDelete:
    properties:
      id:
        description: 编号
        type: integer
    type: object
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  response.PageResult:
    properties:
      list:
        type: object
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data:
        type: object
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /app/forumPosts/createCommentThumbsUp:
    post:
      consumes:
      - application/json
      parameters:
      - description: 评论点赞
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/community.CommentThumbsUp'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 评论点赞
      tags:
      - App_ForumPosts
  /app/forumPosts/createForumComment:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建ForumComment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.CreateForumComment'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建ForumComment
      tags:
      - App_ForumPosts
  /app/forumPosts/createForumPosts:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建ForumPosts
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/community.ForumPosts'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建ForumPosts
      tags:
      - App_ForumPosts
  /app/forumPosts/createForumThumbsUp:
    post:
      consumes:
      - application/json
      parameters:
      - description: 帖子点赞
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/community.ForumThumbsUp'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 帖子点赞
      tags:
      - App_ForumPosts
  /app/forumPosts/deleteCommentThumbsUp:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 取消评论点赞
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteCommentThumbsUp'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 取消评论点赞
      tags:
      - App_ForumPosts
  /app/forumPosts/deleteForumComment:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除ForumComment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdDelete'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除ForumComment
      tags:
      - App_ForumPosts
  /app/forumPosts/deleteForumCommentByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除ForumComment
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除ForumComment
      tags:
      - App_ForumPosts
  /app/forumPosts/deleteForumPosts:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除ForumPosts
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdDelete'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除ForumPosts
      tags:
      - App_ForumPosts
  /app/forumPosts/deleteForumPostsByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除ForumPosts
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除ForumPosts
      tags:
      - App_ForumPosts
  /app/forumPosts/deleteForumThumbsUp:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 取消帖子点赞
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.DeleteForumThumbsUp'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 取消帖子点赞
      tags:
      - App_ForumPosts
  /app/forumPosts/deleteSelfForumPosts:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除自己的ForumPosts
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdDelete'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除自己的ForumPosts
      tags:
      - App_ForumPosts
  /app/forumPosts/findForumComment:
    get:
      consumes:
      - application/json
      parameters:
      - description: 编号
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumComment
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/community.ForumComment'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询ForumComment
      tags:
      - App_ForumPosts
  /app/forumPosts/findForumPosts:
    get:
      consumes:
      - application/json
      parameters:
      - description: 编号
        in: query
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumPosts
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/community.ForumPosts'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 用id查询ForumPosts
      tags:
      - App_ForumPosts
  /app/forumPosts/getForumCommentList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 父评论编号
        in: query
        name: parentId
        type: integer
      - description: 帖子编号
        in: query
        name: postsId
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumComment
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.ForumComment'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取ForumComment列表
      tags:
      - App_ForumPosts
  /app/forumPosts/getForumPostsList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 类别（0视频、1动态、2资讯、3公告、4文章、5问答、6建议）
        in: query
        name: category
        type: integer
      - description: 圈子_编号
        in: query
        name: circleId
        type: integer
      - description: 创建时间（结束）
        in: query
        name: endCreatedAt
        type: string
      - description: 帖子分类编号
        in: query
        name: groupId
        type: integer
      - description: 关键字
        in: query
        name: keyword
        type: string
      - description: 精华(0否 1是)
        in: query
        name: marrow
        type: integer
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - description: 创建时间（开始）
        in: query
        name: startCreatedAt
        type: string
      - description: 标题
        in: query
        name: title
        type: string
      - description: 置顶(0否 1是)
        in: query
        name: top
        type: integer
      - description: 发布者编号
        in: query
        name: userId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 返回community.ForumPosts
          schema:
            allOf:
            - $ref: '#/definitions/response.PageResult'
            - properties:
                List:
                  items:
                    $ref: '#/definitions/community.ForumPosts'
                  type: array
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 分页获取ForumPosts列表
      tags:
      - App_ForumPosts
  /app/forumPosts/updateForumPosts:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新ForumPosts
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/community.ForumPosts'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新ForumPosts
      tags:
      - App_ForumPosts
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"